+++ commands/spigot-platform/src/test/java/net/apartium/cocoabeans/commands/spigot/game/Stage.java
@@ -1,0 +1,11 @@
+package net.apartium.cocoabeans.commands.spigot.game;
+
+public enum Stage {
+
+    WAITING_FOR_PLAYERS,
+    COUNTDOWN,
+    ON_GOING,
+    ENDING,
+    END
+
+}